+++ lib/bibtex.dart
@@ -4,3 +4,4 @@ library;
 
 export 'src/bibtex/definition.dart';
 export 'src/bibtex/model.dart';
+export 'src/bibtex/normalize.dart';